move sorted topics code out of recent
